Fix linker and compiler errors
