+++ src/pipeline/datatype/ImgDetectionsT.cpp
@@ -72,92 +72,6 @@ std::optional<dai::ImgFrame> ImgDetectionsT<DetectionT>::getSegmentationMask() c
     return img;
 }
 
-#ifdef DEPTHAI_HAVE_OPENCV_SUPPORT
-
-template <class DetectionT>
-void ImgDetectionsT<DetectionT>::setCvSegmentationMask(cv::Mat mask) {
-    if(mask.type() != CV_8UC1) {
-        throw std::runtime_error("SetCvSegmentationMask: Mask must be of INT8 type, got opencv type " + cv::typeToString(mask.type()) + ".");
-    }
-    std::vector<std::uint8_t> dataVec;
-    if(!mask.isContinuous()) {
-        for(int i = 0; i < mask.rows; i++) {
-            dataVec.insert(dataVec.end(), mask.ptr(i), mask.ptr(i) + mask.cols * mask.elemSize());
-        }
-    } else {
-        dataVec.insert(dataVec.begin(), mask.datastart, mask.dataend);
-    }
-    setData(std::move(dataVec));  // Call the rvalue overload to allocate a new memory holder
-    this->segmentationMaskWidth = mask.cols;
-    this->segmentationMaskHeight = mask.rows;
-}
-
-template <class DetectionT>
-std::optional<cv::Mat> ImgDetectionsT<DetectionT>::getCvSegmentationMask(cv::MatAllocator* allocator) {
-    if(data->getData().data() == nullptr) {
-        return std::nullopt;
-    }
-    cv::Size size(getSegmentationMaskWidth(), getSegmentationMaskHeight());
-    int type = CV_8UC1;
-    if(size.width <= 0 || size.height <= 0) {
-        throw std::runtime_error("Segmentation mask metadata not valid (width or height <= 0).");
-    }
-
-    const size_t requiredSize = CV_ELEM_SIZE(type) * static_cast<size_t>(size.area());
-    const size_t actualSize = data->getSize();
-
-    if(actualSize != requiredSize) {
-        throw std::runtime_error("Segmentation mask data size does not match the expected size, required " + std::to_string(requiredSize) + ", actual "
-                                 + std::to_string(actualSize) + ".");
-    }
-
-    cv::Mat mask;
-    mask = cv::Mat(size, type, data->getData().data());
-    CV_Assert(mask.type() == CV_8UC1);
-
-    cv::Mat output;
-    if(allocator != nullptr) {
-        output.allocator = allocator;
-    }
-    (mask).copyTo(output);
-    return output;
-}
-
-template <class DetectionT>
-std::optional<cv::Mat> ImgDetectionsT<DetectionT>::getCvSegmentationMaskByIndex(uint8_t index, cv::MatAllocator* allocator) {
-    std::optional<cv::Mat> mask = getCvSegmentationMask(allocator);
-    if(!mask.has_value()) {
-        return std::nullopt;
-    }
-    cv::Mat classMask;
-    cv::compare(*mask, index, classMask, cv::CmpTypes::CMP_EQ);
-
-    return classMask;
-}
-
-template <class DetectionT>
-std::optional<cv::Mat> ImgDetectionsT<DetectionT>::getCvSegmentationMaskByClass(uint8_t semanticClass, cv::MatAllocator* allocator) {
-    std::optional<cv::Mat> mask = getCvSegmentationMask(allocator);
-    if(!mask.has_value()) {
-        return std::nullopt;
-    }
-    cv::Mat classMask = cv::Mat::zeros((*mask).size(), CV_8UC1) + 255;
-
-    for(uint8_t idx = 0; idx < detections.size(); idx++) {
-        if(detections[idx].label == semanticClass) {
-            std::optional<cv::Mat> indexMask = getCvSegmentationMaskByIndex(idx, allocator);
-            if(!indexMask.has_value()) {
-                return std::nullopt;
-            }
-            classMask.setTo(0, *indexMask);
-        }
-    }
-
-    return classMask;
-}
-
-#endif
-
 template class ImgDetectionsT<dai::ImgDetection>;
 template class ImgDetectionsT<dai::SpatialImgDetection>;
