+++ src/utility/RecordReplay.cpp
@@ -234,12 +234,5 @@ std::string matchTo(const std::vector<std::string>& deviceIds, const std::vector
     return deviceId;
 }
 
-#ifndef DEPTHAI_HAVE_OPENCV_SUPPORT
-std::tuple<size_t, size_t, double> getVideoSize(const std::string& filePath) {
-    (void)filePath;
-    throw std::runtime_error("OpenCV is required to get video size");
-}
-#endif
-
 }  // namespace utility
 }  // namespace dai
